Daily Update, 2-13-22
added the basics.html page, this page is ready for css formatting.

photoshoot page is for wirting out small snipets of code you need to take a screenshot of for reference.

hello world is again a sample thing for a screenshot,

tag list is incomplete but in progress.

style.css has been created and linked to some of the pages

relevent photos for examples have been added. migt need to consider putting them into  a new folder if possible to keep the main folder frim getting too cluttered,