2.0.1
fixed some minor sintax issues and added in the globalstyle css page to play with later. also this marks the official start of the new naming convention (num1).(num2).(num3) where num1= the major build version, this only changes is we do a major refactor or structure change. something that requires changing the entire directory structure or we impamet a mjor design change that affects every page. num2= the minor version. basically anytime we merge the develop branch with the main. num3= the feature version, anytime we merge in a new feature branch this number will go up by 1. since the main has already udergone a major file structure and formatting refactor we are on version 2.x.x and since this is the first feature branch to be merged up since starting this naming convention i am starting with 2.0.1